make 64-bit build work
Characters can be created, saved, and loaded. Playing seems to work. Not
tested much beyond that.

+++ chr_file.c
@@ -120,7 +120,7 @@ int unlock;
 #endif
     return(NOPE);
   }
-  tmp = sizeof(long) + (numc - numc2 - 1) * sizeof(struct chr);
+  tmp = sizeof(int) + (numc - numc2 - 1) * sizeof(struct chr);
   lseek(cwfp, tmp, L_SET);
   strncpy(sav.nam[0], u.n[0], NAMELEN);
   strncpy(sav.nam[1], u.n[1], NAMELEN);
@@ -169,7 +169,7 @@ int lock;
     printf("[File busy - please wait]\r\n");
     sleep(roll(1,2));
   }
-  loc = sizeof(long) + (numc - numc2 - 1) *sizeof(struct chr);
+  loc = sizeof(int) + (numc - numc2 - 1) *sizeof(struct chr);
   lseek(cwfp, loc, L_SET);
   if (lock == LOCK) {
     read(cwfp, (char *) &zold, sizeof(struct chr));
@@ -217,14 +217,14 @@ int chr_new()      /* save u. in file, lock it */
     if (cp->c[0] == 0)
       break;
   if (cp != NULL) {
-    tmp = sizeof(long) + (numc - numc2 - 1) * sizeof(struct chr);
+    tmp = sizeof(int) + (numc - numc2 - 1) * sizeof(struct chr);
     lseek(cwfp, tmp, L_SET);
   } else {
     numc++;
     tmp = htonl(numc);
     lseek(cwfp, 0, L_SET);
     write(cwfp, (char *) &tmp, sizeof(tmp));
-    tmp = sizeof(long) + (numc - 1) * sizeof(struct chr);
+    tmp = sizeof(int) + (numc - 1) * sizeof(struct chr);
     lseek(cwfp, tmp, L_SET);
   } 
   strncpy(cpout.nam[0], u.n[0], NAMELEN);
